update the validation schema for YAML metadata file, refactor code
