+++ src/syngen/ml/validation_schema/__init__.py
@@ -1,1 +1,2 @@
 from syngen.ml.validation_schema.validation_schema import validate_schema
+from syngen.ml.validation_schema.validation_schema import configuration_schema
